requirements.txt
SQLAlchemy~=2.0.23
PyJWT~=2.8.0
environs~=9.5.0
fastapi~=0.105.0
pydantic~=2.5.2
passlib~=1.7.4
python-jose~=3.3.0
